LPS-179177 Extract CommerceInventoryBookedScreenNavigationCategory methods to a CommerceInventoryBookedScreenNavigationEntry component and extends CommerceInventoryBookedScreenNavigationCategory in it
