+++ modules/apps/commerce/commerce-inventory-web/src/main/java/com/liferay/commerce/inventory/web/internal/frontend/taglib/servlet/taglib/CommerceInventoryBookedScreenNavigationCategory.java
@@ -14,21 +14,13 @@
 
 package com.liferay.commerce.inventory.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.inventory.model.CommerceInventoryReplenishmentItem;
 import com.liferay.commerce.inventory.web.internal.constants.CommerceInventoryScreenNavigationConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
-import java.io.IOException;
-
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -37,15 +29,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Luca Pellizzon
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceInventoryBookedScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommerceInventoryReplenishmentItem> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -54,16 +42,11 @@ public class CommerceInventoryBookedScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return getCategoryKey();
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, getCategoryKey());
+		return language.get(resourceBundle, getCategoryKey());
 	}
 
 	@Override
@@ -72,20 +55,7 @@ public class CommerceInventoryBookedScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_INVENTORY;
 	}
 
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse, "/details/booked.jsp");
-	}
-
 	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
+	protected Language language;
 
 }
